doom_ddqn/requirements.txt
--extra-index-url https://download.pytorch.org/whl/cpu
torch
torchvision
numpy
vizdoom
pillow
matplotlib
flask